backend/requirements.txt
fastapi[standard]==0.136.0
uvicorn[standard]==0.44.0
python-multipart==0.0.26
langchain==1.2.15
langchain-upstage==0.7.7
pillow==12.2.0
pdf2image==1.17.0
python-dotenv==1.2.2
